13/09/23 MAJ2

+++ view/forum/listTopics.php
@@ -11,27 +11,29 @@ $category = $result["data"]['category'];
 <h2>Listes des topics</h2>
 
 <?php
-foreach($topics as $topic){
-?>
-    <div class="d-flex flex-row gap-5">
-        <p>
-            <a href="index.php?ctrl=forum&action=findMessageByTopicId&id=<?=$topic->getId();?>"><?= $topic->getNameTopic() ?></a>
-        </p>
-        <p>De : <?= $topic->getUser()->getUserName() ?></p>
-        <p>Le : <?= $topic->getDateCreationTopic() ?></p>
-        <!-- UPDATE -->
-        <!-- Quand on clique on récupère idTopic et categoryId pour être sûr de changer le topic dans la bonne catégorie :
-        $topic->getId() ?>&categoryId=< $category->getId() ?>
-        -->
-        <a href="index.php?ctrl=forum&action=updateTopicForm&id=<?= $topic->getId() ?>&categoryId=<?= $category->getId() ?>">
-            <i class="fa-solid fa-pencil" title="Update"></i>
-        </a>
-        <!-- DELETE -->
-        <a href="index.php?ctrl=forum&action=deleteTopic&id=<?= $topic->getId() ?>&categoryId=<?= $category->getId() ?>">
-            <i class="fa-regular fa-trash-can" title="Delete"></i>
-        </a>
-    </div>
-<?php
+if ($topics !== NULL) {
+    foreach($topics as $topic){
+    ?>
+        <div class="d-flex flex-row gap-5">
+            <p>
+                <a href="index.php?ctrl=forum&action=findMessageByTopicId&id=<?=$topic->getId();?>"><?= $topic->getNameTopic() ?></a>
+            </p>
+            <p>De : <?= $topic->getUser()->getUserName() ?></p>
+            <p>Le : <?= $topic->getDateCreationTopic() ?></p>
+            <!-- UPDATE -->
+            <!-- Quand on clique on récupère idTopic et categoryId pour être sûr de changer le topic dans la bonne catégorie :
+            $topic->getId() ?>&categoryId=< $category->getId() ?>
+            -->
+            <a href="index.php?ctrl=forum&action=updateTopicForm&id=<?= $topic->getId() ?>&categoryId=<?= $category->getId() ?>">
+                <i class="fa-solid fa-pencil" title="Update"></i>
+            </a>
+            <!-- DELETE -->
+            <a href="index.php?ctrl=forum&action=deleteTopic&id=<?= $topic->getId() ?>&categoryId=<?= $category->getId() ?>">
+                <i class="fa-regular fa-trash-can" title="Delete"></i>
+            </a>
+        </div>
+    <?php
+    }
 }
 ?>
